# docs/site/openapi.yaml
...
openapi: 3.0.3
info:
  title: Regente — integration API
  version: "1.0"
  description: |
    **Curated** contract of Regente's integration surface: the routes an
    external system actually calls. The SPA's internal surface (Design
    sessions, users, settings, alerting...) is deliberately left out — it
    evolves with the UI and is not a contract.

    ### Authentication
    The security declaration below models the **integration bearer transport**.
    In `local` or `hybrid` mode, POST `/api/auth/login` with
    `{"username":"...","password":"..."}` (omit `browser`, or set it false)
    returns `{token,user}`. Send that token as `Authorization: Bearer <token>`.
    The deployment's static `REGENTE_TOKEN` is also accepted as admin in those
    modes. Neither ordinary password login nor the static token bypasses `oidc`
    mode. Explicitly configured emergency access is a separate, audited,
    15-minute recovery path, not a normal integration credential.

    **Browser sessions are different:** login with `browser:true` returns an
    empty JSON token and an HttpOnly, SameSite=Lax cookie (`regente_session` on
    HTTP, `__Host-regente_session` with Secure on HTTPS). OIDC also uses cookies.
    Cookie requests require an allowed origin and, except GET/HEAD/OPTIONS,
    `X-CSRF-Token` from login or GET `/api/auth/me` (including POST/QUERY reads).
    Never copy browser cookies into bearer headers, URLs or localStorage.
...

...

...

...

    ### RBAC
...

    ### The QUERY HTTP verb
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/instances/query:
    post:
      tags: [Query]
      summary: Typed composed query (D-5)
      description: |
        Composed filtering beyond the query string of `/api/instances`: date
        ranges, IN lists (folders/statuses), flags (forced/carried/late) and
        aggregation (`groupBy`). STRICT parsing — an unknown field is a 400
        (this is not SQL-over-HTTP; every filter is typed and parameterized).
        An empty body `{}` means today's daily, unfiltered. Keyset pagination
        via `cursor` = the id of the last item on the previous page (only
        without `groupBy`).

        Also accepts the `QUERY` HTTP method (same handler/contract).
      operationId: queryInstances
      x-permission: any logged-in user (reads filtered by folder ACLs)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StructuredQuery" }
            example: { statuses: [NOTOK], limit: 100 }
      responses:
        "200":
          description: Paginated list (without groupBy) OR aggregation (with groupBy).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/QueryResultList"
                  - $ref: "#/components/schemas/QueryResultGrouped"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/query:
    post:
      tags: [Query]
      summary: NL-query (text → structured query)
      description: |
        Interprets a natural-language question (en/pt) about the current day —
        "how many jobs failed today?", "why didn't etl-sales run?" — and runs
        the intent (summary/count/list/explain) against the database.
        Deterministic (an intent parser, no AI). The response returns the
        interpreted intent plus the structured answer.
      operationId: nlQuery
      x-permission: any logged-in user (reads filtered by folder ACLs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [q]
              properties:
                q: { type: string, description: "The question, in natural language." }
            example: { q: "how many jobs failed today?" }
      responses:
        "200":
          description: Interpreted intent + answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  interpreted:
                    type: object
                    description: How the question was understood (kind/status/folder/job).
                  answer:
                    type: object
                    description: Structured answer (its shape varies with the kind — text/count/byStatus/items...).
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/forecast:
    get:
      tags: [Forecast]
      summary: Forecast for one day (F21)
      description: |
        Simulates in RAM what the daily would materialize on `date`: who is
        eligible (the SAME gating as RunDaily — frequency, include/exclude
        calendars, shift, via IsScheduledOn), topological order (waves), blocks
        from deps/resources/conditions and projected SLA violations. It never
        touches the database.
      operationId: getForecast
      x-permission: any logged-in user
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
          description: YYYY-MM-DD; defaults to today.
      responses:
        "200":
          description: Simulation of the day.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForecastReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/forecast/range:
    get:
      tags: [Forecast]
      summary: Forecast for N days ahead (CTM-6)
      description: One ForecastReport per day, same gating as RunDaily. The classic "≥1 week".
      operationId: getForecastRange
      x-permission: any logged-in user
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
          description: YYYY-MM-DD; defaults to today.
        - name: days
          in: query
          schema: { type: integer, default: 7, minimum: 1, maximum: 366 }
      responses:
        "200":
          description: One report per day.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ForecastReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Non-browser API login session, or static REGENTE_TOKEN (= admin) in local/hybrid mode. Not an HttpOnly browser session, OIDC provider token or agent machine credential. Static tokens do not bypass oidc mode.
...
  responses:
    BadRequest:
      description: Invalid request (raw message in text/plain).
      content:
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: Missing/invalid credentials or session source rejected by the authentication mode.
      content:
        text/plain:
          schema: { type: string, example: unauthorized }
...
  schemas:
    BusinessTime:
      type: object
      required: [timezone, dailyAt]
      properties:
        timezone: { type: string, description: Frozen IANA timezone. }
        dailyAt: { type: string, description: Frozen rollover in HH:MM. }
    Instance:
      type: object
      description: |
        One instance in the daily (an IMMUTABLE snapshot of the order —
        team/dryRun and friends freeze at the moment the job is ordered).
      properties:
        businessTime: { $ref: "#/components/schemas/BusinessTime" }
        id: { type: string }
        definitionId: { type: string }
        team: { type: string, description: The folder (frozen on the order). }
        orderDate: { type: string, format: date }
        status:
          type: string
          enum: [WAITING, RUNNING, OK, NOTOK, HELD, CANCELLED]
        scheduledAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        agentId: { type: string }
        exitCode: { type: integer }
        output: { type: string }
        forced: { type: boolean }
        carriedFrom:
          type: string
          format: date
          description: The day of origin if it came from another daily's carry-over.
        confirmed: { type: boolean, description: "Confirm gate opened (jobs with confirm: true)." }
        cycleRuns: { type: integer, description: OK laps completed during the day (cyclic jobs). }
        dryRun: { type: boolean, description: Frozen on the order (the GHOST badge). }
...
    StructuredQuery:
      type: object
      description: Everything optional; the zero value means today's daily, unfiltered. An unknown field is a 400.
      additionalProperties: false
      properties:
        date: { type: string, format: date, description: An exact day (YYYY-MM-DD). }
        from: { type: string, format: date, description: ...or an inclusive from..to range. }
        to: { type: string, format: date }
        folders:
          type: array
          items: { type: string }
          description: IN over the folder/team.
        statuses:
          type: array
          items: { type: string }
          description: IN over the status (WAITING/RUNNING/OK/NOTOK/HELD/CANCELLED).
        search: { type: string, description: LIKE over id / definitionId. }
        defIdPrefix: { type: string, description: definitionId LIKE 'prefix%'. }
        forced: { type: boolean, description: Only forced ones (true) / only non-forced ones (false). }
        carried: { type: boolean, description: Only carry-overs from another daily (true) / only native ones (false). }
        late: { type: boolean, description: "true = WAITING with scheduledAt in the past (late)." }
        groupBy:
          type: string
          enum: [status, folder, definition]
          description: Aggregation in the database — switches the response to the grouped shape.
        limit: { type: integer, default: 500, maximum: 5000 }
        cursor: { type: string, description: Keyset — the id of the last item on the previous page. }
    QueryResultList:
      type: object
      description: The paginated shape (without groupBy).
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Instance" }
        nextCursor:
          type: string
          description: Pass it back in `cursor` for the next page; empty means there is nothing left.
    QueryResultGrouped:
      type: object
      description: The aggregated shape (with groupBy) — counters straight from the database, no row materialized.
      properties:
        groupBy: { type: string }
        groups:
          type: array
          items:
            type: object
            properties:
              key: { type: string }
              count: { type: integer }
        total: { type: integer }
...
    ForecastReport:
      type: object
      properties:
        businessTime: { $ref: "#/components/schemas/BusinessTime" }
        orderDate: { type: string, format: date }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/ForecastJob" }
        peakResourceUsage:
          type: object
          additionalProperties: { type: integer }
          description: Projected peak usage per resource.
    ForecastJob:
      type: object
      properties:
        defId: { type: string }
        label: { type: string }
        team: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        eligible: { type: boolean }
        reason: { type: string, description: 'Why it does not run ("no calendar match", "blocked by deps"...).' }
        wave: { type: integer, description: Topological order (0 = root). }
        wouldBreachSla: { type: boolean }
...